[lab04] run.sh update

+++ lab04/algs.h
@@ -20,10 +20,10 @@ struct bucket {
 
 int cmp(const void *a, const void *b);
 
-int sort_v1(item_t* array, int32_t arr_size, int16_t n_buckets, struct bucket *buckets,
+int sort_v1(item_t* array, int32_t arr_size, int32_t n_buckets, struct bucket *buckets,
         unsigned int thread_idx, int8_t num_of_threads);
 
-int sort_v2(item_t *array, int32_t arr_size, int16_t n_buckets, struct bucket *buckets);
+int sort_v2(item_t *array, int32_t arr_size, int32_t n_buckets, struct bucket *buckets);
 
 int rand_arr(item_t* array, int32_t arr_size, u_int tid);
