+++ lab04/params.h
@@ -9,8 +9,8 @@
 
 typedef struct param {
     int8_t alg;
-    int8_t threads;
-    int16_t n_buckets;
+    int32_t threads;
+    int32_t n_buckets;
     int32_t arr_size;
 } param;
